requirements.txt
colorama
discord.py
pystyle
requests
hmtai
discord.py-self